Reimplement HousingValidation
